arreglos

+++ pet/src/main/java/com/pethistory/pet/dtos/DtoUsuarioGett.java
@@ -26,7 +26,7 @@ public class DtoUsuarioGett {
 
     private Date fechaNac;
 
-    private Integer tipdoc;
+    private Long tipdoc;
 
     private String nomTipoDoc;
